More fixups for Python 3.10
For some reason now, PyQt5 doesn't auto-coerce args from int -> float
and float -> int, so everything that is passing the wrong type from
Python to PyQt is broken now.  Luckily this doesn't happen in too many
places, but it does happen.

This commit fixes more of these errors leading to app crash.

+++ electroncash_gui/qt/utils/aspect_svg_widget.py
@@ -37,5 +37,5 @@ class FixedAspectRatioSvgWidget(QSvgWidget):
     def sizeHint(self) -> QSize:
         svg_size = super().sizeHint()
         aspect_ratio = svg_size.width() / svg_size.height()
-        size_hint = QSize(self._width, self._width / aspect_ratio)
+        size_hint = QSize(int(self._width), int(self._width / aspect_ratio))
         return size_hint
